Handle complete, but not unique input for completion in a sane way and apply multiple patches by Brent Goodrick:
Add an additional level of indirection for direct access to gimp-oblist-cache to instead use a new function called gimp-get-oblist-cache that restores it from the file on demand.  There was a case in one of the tests (I don't recall which) that ran afoul of a nil gimp-oblist-cache without it.

gimp-save-input-ring needs to check that there is an actual gimp-proc before attempting to use the buffer.

gimp-unit-test now calls gimp-quit at the very end, so that I could run the unit test in a tight loop to validate that the hang didn't show up again.

Don't allow gimp-script-p to fail when there is no file associated with the current buffer.  This can occur for gimp-mode buffers that are not associated with files (shows up because gimp-script-p is in magic-mode-alist).

Fix one case of hanging in `gimp-first-run-action'. Fix troubles in unit tests.